src/app/rxjs/hot-stream/hot-stream.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { HotStreamComponent } from './hot-stream.component';

describe('Component: HotStream', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [HotStreamComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([HotStreamComponent],
      (component: HotStreamComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(HotStreamComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(HotStreamComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-hot-stream></app-hot-stream>
  `,
  directives: [HotStreamComponent]
})
class HotStreamComponentTestController {
}

